demos/inferences/requirements.txt
requests>=0.24.0
paho-mqtt>=1.6.0
